ENH: Add Python Jupyter Notebook testing procedure
Adds job to `build-test-package-python` workflow that retrieves a built
Linux Python wheel artifact previously output by the workflow and uses
it to test Jupyter Notebooks in the `examples/` folder of the external
module.

For initial implementation only Linux wheels are tested with the Python
version fixed at 3.10. To coincide with the Github Actions Linux runner
image and machine architecture, only the `_2_28-x64` wheel is used.

Closes
https://github.com/InsightSoftwareConsortium/ITKRemoteModuleBuildTestPackageAction/issues/49
.