ENH: Bump CMake and checkout GitHub actions versions
Bump CMake and checkout GitHub actions versions.

Fixes:
```
cxx-build-workflow / build-test-cxx (ubuntu-22.04)
The following actions uses Node.js version which is deprecated and will be forced to run on node20:
lukka/get-cmake@v3.24.2.
For more info:
https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
```

and
```
python-build-workflow / build-macos-py (8)
The following actions uses Node.js version which is deprecated and will be forced to run on node20:
actions/checkout@v3.
For more info:
https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
```

raised for example in:
https://github.com/InsightSoftwareConsortium/ITKAnalyzeObjectMap/actions/runs/10012233814